Some random shit
Tried fucking around with asteroids, didn't get them in. Added a crappy
kick to the server if it doesn't get anything from a client after a
certain amount of time. Similar thing is on the client but doesn't work
when they're by them self because the server doesn't send anything in
that case. Needs asteroids or something.